Preview playing, fix success rate bar.

+++ osu.Game/Screens/Select/Details/SuccessRate.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
 using OpenTK;
-using OpenTK.Graphics;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -10,12 +9,7 @@ using osu.Game.Database;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
-using System.Globalization;
 using System.Linq;
-using osu.Game.Online.API;
-using osu.Game.Online.API.Requests;
-using osu.Framework.Threading;
-using osu.Framework.Graphics.Shapes;
 
 namespace osu.Game.Screens.Select.Details
 {
@@ -37,7 +31,7 @@ namespace osu.Game.Screens.Select.Details
                 beatmap = value;
 
                 successPercent.Text = $"{beatmap.OnlineInfo.SuccessRate}%";
-                successRate.Length = beatmap.OnlineInfo.SuccessRate / 100;
+                successRate.Length = (float)beatmap.OnlineInfo.SuccessRate / 100f;
 
                 var retries = Beatmap.Metrics.Retries;
                 var fails = Beatmap.Metrics.Fails;
